Starting Point for code
All code has comments and story board is organized. view controller
files are in file associated with the tab they belong to. Need to
finish user profile so we can start sending friends request. the
friends button on the main profile just contains a dummy array, not
actually users. also need to create a bar activity in firebase when the
user changes his or hers current bar. this will help when creating a
news feed, but we should optimize what we have now before adding more
complexity to the code.