added framework and users profiles
add a profile for users when you search for them. you can now view
other users friends. still need to add the ability to add a real
friend. added a framework for messaging, but will deal with that later,
just found it when googling and thought it would be cool.